+++ changelog.md
@@ -5,6 +5,7 @@ Version 2.5.0
 - Updated `cardinal-components-level` and `cardinal-components-chunk` to 1.16.2
 
 Additions
+- Added a new API for block components (see PR #42)
 - Added reworked `ComponentProvider` and `ComponentContainer` interfaces to the V3 API
     - Refactored a lot of internal code to use the new types
 - Added `ComponentV3` and `ComponentRegistryV3` interfaces for early V3 adopters